% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackNodeRoutes.R
\name{blackNodeRoutes}
\alias{blackNodeRoutes}
\title{blackNodeRoutes}
\usage{
blackNodeRoutes(rank, nodePosition)
}
\arguments{
\item{rank}{This is the Rank of the maze.}

\item{nodePosition}{tells you all the position of the black dots}
}
\description{
This function returns the frequently of paths going through a sequence of black dots
}
\details{
This function returns the frequently of paths going through a sequence of black dots
}
\author{
Aiden Loe and Maria
}

