% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{cal}
\alias{cal}
\title{Calculate}
\usage{
cal(rank, nodePosition)
}
\arguments{
\item{rank}{This is the Rank of the maze.}

\item{nodePosition}{Tells you all the position of the black dots.}
}
\description{
This returns several results.
}
\details{
This function calculates the count of all the possible black node routes, the maximum score one can achieve for a given rank of a colour node position, all the minimum routes possible, and all the possible routes.
}
\examples{
rank <- 10
nodePosition <- colourNodePosition(rank=10,satPercent=0.5,seed=16)
c <- cal(rank,nodePosition)
}
\author{
Aiden Loe
}

