% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colleage.R
\name{colleage}
\alias{colleage}
\title{colleage}
\usage{
colleage(n, rank)
}
\arguments{
\item{n}{This is the node number.}

\item{rank}{This is the Rank of the maze.}
}
\description{
This function tells us which nodes are connected to the nodes.
}
\details{
This function tells us which nodes are connected to the node you are interested in based on a given Grid.
}
\author{
Aiden Loe
}

