% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmaze.R
\name{genMaze}
\alias{genMaze}
\title{genMaze}
\usage{
genMaze(rank)
}
\arguments{
\item{rank}{This is the Rank of the maze.}
}
\description{
This function generates the list of edges.
}
\details{
The Genmaze function generates the list of edges. The edges will be used to construct the maze.
}
\author{
Aiden Loe
}

