% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_model_8.R
\name{imEight}
\alias{imEight}
\title{Item Model 8}
\usage{
imEight(cat, n, items, arith)
}
\arguments{
\item{cat}{Number of categorical groups per question.}

\item{n}{The differences between the pair of objects}

\item{items}{The number of items you want to generate.}

\item{arith}{The arithmetic operator of your choice ("add","substr","multi","div").}
}
\description{
This uses item model 8 to create number series items.
}
\details{
This is based on the categorical / pattern recognition rule. Neighbouring pairs or triads of objects are related, includes arithmetic operations.
}
\examples{

imEight(cat=2,n=4,items=2, arith="add")


}
\author{
Aiden Loe and Filip Simonfy
}
