% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_model_5.R
\name{imFive}
\alias{imFive}
\title{Item Model 5}
\usage{
imFive(arithOne = "add", arithTwo = "substr", n = 2, items = 4)
}
\arguments{
\item{arithOne}{Select the arithmetric operator of choice ("add","multi", "sub", "div").}

\item{arithTwo}{Select the arithmetric operator of choice ("add","multi", "sub", "div").}

\item{n}{Value you want use the arithmetic operator on.}

\item{items}{Generate a random mix of items.}
}
\description{
This uses item model 5 to create number series items - Identification of co-occurring relationships between elements (with use of arithmetic skills)
}
\details{
Logic analogous to the Item Model 4, but at least one sub-sequence involves the basic arithmetic operations. Sequences combine items from Item Families 1 and 3. The arithmetic operations change but the differences in value remains the name. Example: Odd elements of the sequence increase by 2 and even elements of the sequence are multiplied by 2. (2 12 4 24 6 48 8 (96) (10))
}
\examples{

imFive(arithOne="add",arithTwo="add",n=2,items=5)


}
\author{
Aiden Loe and Filip Simonfy
}
