% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_model_4.R
\name{imFour}
\alias{imFour}
\title{Item Model 4}
\usage{
imFour(items = 5, seed = 1)
}
\arguments{
\item{items}{Number of items to generate.}

\item{seed}{This gives you the same result again.}
}
\description{
This uses item model 4 to create number series items - Identification of co-occurring relationships between elements (without use of arithmetic skills)
}
\details{
Sequences which consist of regularly alternating parallel sub-sequences. Understanding of succession does not require use of algebraic skill. Sub-sequences involve items from Item Model 1. Example: Odd elements of the sequence are multiples of 1 and even elements of the sequence are multiples of 10. (1 10 2 20 3 30 (4) (40))
2 simple linear (without arithmetic) 1 2 3 / 10 20 30 combine to form a number series item.
}
\examples{
\dontrun{

imFour(items=5, seed=5)

}
}
\author{
Aiden Loe and Filip Simonfy
}
