% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_model_9.R
\name{imNine}
\alias{imNine}
\title{Item Model 9}
\usage{
imNine(items)
}
\arguments{
\item{items}{Number of items to generate.}
}
\description{
This uses item model 10 to create number series items - Identification of relationships within a chain of elements.
}
\details{
Progressive sequences which involve relationships between multiple preceding objects (e.g. Fibonacci sequence). Example: Each element of the sequence is a result of addition of its two preceding elements (1 1 2 3 5 8 (13)). The maximum number to be generated is 15 items.
}
\examples{

imNine(items=3)


}
\author{
Aiden Loe and Filip Simonfy
}
