% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_model_1.R
\name{imOne}
\alias{imOne}
\title{Item Model 1}
\usage{
imOne(items = 5, seed = 1)
}
\arguments{
\item{items}{Number of items to generate.}

\item{seed}{Setting the seed returns the same items on the local computer.}
}
\description{
This uses item model 1 to create number series items - Elementary understanding of sequence succession.
}
\details{
Simple linear sequences which do not require use of advanced arithmetic operations, such as ordered multiples of 1, 10, or 100. Example: A sequence of ordered multiples of 10. (10 20 30 40 (50)).
}
\examples{

imOne(items=5, seed=5)


}
\author{
Aiden Loe and Filip Simonfy
}
