% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_model_6.r
\name{imSix}
\alias{imSix}
\title{Item Model 6}
\usage{
imSix(items)
}
\arguments{
\item{items}{Number of items to generate.}
}
\description{
This uses item model 6 to create number series items - Identification of progressively evolving coefficients of change.
}
\details{
Non-linear progressive sequences which require a higher level of abstraction; the coefficient of change between two neighbouring elements is not invariable and its elements form a sequence. The coefficient sequences correspond to items from Item Families 1 and 3. Example: The coefficient of change between each pair of neighbouring elements in the sequence increases by 1. (2 4 7 11 16 (22))
}
\examples{

imSix(items=3)

}
\author{
Aiden Loe and Filip Simonfy
}
