% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_model_13.R
\name{imThirteen}
\alias{imThirteen}
\title{Item Model 13}
\usage{
imThirteen(items)
}
\arguments{
\item{items}{Generate a random mix of items.}
}
\description{
This uses item model 13 to create number series items - Combined identification of unevenly ordered sub-sequences and non-successive relationships between elements.
}
\details{
This function creates number series creates a combination of sequences and ratios. TLogic analogous to the Item Model 13, but the second sequence belongs to the Item Model 9. As a result, pairs of elements following certain rule are embedded into a progressive sequence. Example: Sequence with coefficient of (+ 1) is interposed with pairs of elements which differ by 3. 1 5 8 2 209 212 3 41 (44) (4). Only the addition and substraction arimethic operators are used to generate the number series items.
}
\examples{

#Draws 10 items randomly.
imThirteen(10)


}
\author{
Aiden Loe and Filip Simonfy
}
