% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_model_12.R
\name{imTwelve}
\alias{imTwelve}
\title{Item Model 12}
\usage{
imTwelve(items)
}
\arguments{
\item{items}{Generate a random mix of items.}
}
\description{
This uses item model 12 to create number series items - Identification of unevenly ordered sub-sequences
}
\details{
This function creates number series that is a irregular combination of sequences a b b a b b a ... Only the addition and substraction arithmetic operators are used to create the number series items.
}
\examples{

#Draws 10 items randomly.
imTwelve(10)


}
\author{
Aiden Loe and Filip Simonfy
}
