% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_model_2.R
\name{imTwo}
\alias{imTwo}
\title{Item Model 2}
\usage{
imTwo(cat = 2, items = 4, random = FALSE)
}
\arguments{
\item{cat}{Length of categorical groups per question.}

\item{items}{The number of items you want to generate.}

\item{random}{To randomise the position of the numeric values.}
}
\description{
This uses item model 2 to create number series items - Understanding of object categorisation.
}
\details{
Sequences consist of elements belonging to two homogeneous groups with equal number of elements. Missing element belongs to the group with fewer elements present in the sequence. For example, 1 1 1 5 5 (5).
}
\examples{

imTwo(cat=2,items=4,random=FALSE)


}
\author{
Aiden Loe and Filip Simonfy
}
