% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookUniqueNodePositions.R
\name{lookUniqueNodePositions}
\alias{lookUniqueNodePositions}
\title{lookUniqueNodePositions}
\usage{
lookUniqueNodePositions(rank, satPercent, seed)
}
\arguments{
\item{rank}{This is the Rank of the maze.}

\item{satPercent}{Percentage of saturation.}

\item{seed}{Returns a unique node distribution specific to the local computer.}
}
\description{
The function returns a black nodes distribution with a single unique solution and seed number
}
\details{
The function returns a black nodes distribution with a single unique solution and seed number
}
\examples{
\dontrun{

rank <- 5
satPercent <- 0.5

#Number of unique solutions
lookUniqueSolution(rank,satPercent,11)

#Searches for just one unique solution
justOne <- genUniqueSolution(rank,satPercent,11)

#Black nodes distribution for single unique solution + seed number
lookUniqueNodePositions(justOne)

#Black nodes distribution
nodePosition <- colourNodePosition(rank,satPercent,justOne)

#Compare to the Solution function
solution(rank,satPercent, nodePosition)
}
}
\author{
Aiden Loe and Maria Sanchez
}

