% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowerGrid.R
\name{lowerGrid}
\alias{lowerGrid}
\title{lowerGrid}
\usage{
lowerGrid(R)
}
\arguments{
\item{R}{This is the Rank of the maze.}
}
\description{
This tells you all the nodes position in the lower grid.
}
\details{
returns the nodes positioned at the lower grid
}
\author{
Aiden Loe
}

