% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxScoreRoutes.R
\name{maxScoreRoutes}
\alias{maxScoreRoutes}
\title{maxScoreRoutes}
\usage{
maxScoreRoutes(rank, nodePosition)
}
\arguments{
\item{rank}{This is the Rank of the maze.}

\item{nodePosition}{tells you all the position of the black dots}
}
\description{
This tells you all the possible routes to achieve a maximum score based on the colourNode position and a given rank.
}
\details{
The maxScoreRoutes function tells you the possible routes to achieve a maximum score.
You need to use the nodePosition function first prior to using this.
}
\author{
Aiden Loe and Maria
}

