% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difficulty.R
\name{mazeDiff}
\alias{mazeDiff}
\title{Maze Diffculty}
\usage{
mazeDiff(rank, satPercent, nodePosition)
}
\arguments{
\item{rank}{This is the Rank of the maze.}

\item{satPercent}{This is the saturation}

\item{nodePosition}{This is the distribution of the colour node positions}
}
\description{
This function tells us the difficulty level of the rank given a saturation and black node distribution
}
\details{
This function tells us the difficulty level of the rank given a saturation and black node distribution. The calculation of the difficulty level follows the paper at XXXX. There are however, more than one way to calculate difficulty. We followed this because it was incorporated all the possible parameters of the task features that may potentially influence item difficulty.
}
\examples{
\dontrun{

rank <- 5
satPercent <- 0.5

#Black nodes distribution
nodePosition <- colourNodePosition(rank,satPercent,1)

#calculate difficulty
mazeDiff(rank,satPercent ,nodePosition)
}
}
\author{
Aiden Loe and Maria Sanchez
}

