% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel2_3.R
\name{nmCombo}
\alias{nmCombo}
\title{Item Model 4 and 5}
\usage{
nmCombo(items, seed)
}
\arguments{
\item{items}{Number of items to generate.}

\item{seed}{This gives you the same result again.}
}
\description{
This uses item model 4 and 5 to create number series items.
}
\details{
Comprehension of abstract object representation
i.e combination of:same number / same letter
2 simple linear (without arithmetic) 1 2 3 / a b c / 10 20 30 combine to form a number series item.
}
\examples{
\dontrun{

nmCombo(items=5, seed=5)

}
}
\author{
Aiden Loe and Filip Simonfy
}

