% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel3_5.R
\name{nmFib}
\alias{nmFib}
\title{Item Model 10}
\usage{
nmFib(items)
}
\arguments{
\item{items}{Number of items to generate.}
}
\description{
This function creates Fibonacci sequences. The maximum number to be generated is 15 items.
}
\details{
Fibonacci sequence
}
\examples{
\dontrun{

nmFib(items=3)

}
}
\author{
Aiden Loe and Filip Simonfy
}

