% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nModel2_3.R
\name{nmFive}
\alias{nmFive}
\title{Item Model 5}
\usage{
nmFive(items, seed = 1)
}
\arguments{
\item{items}{Number of items to generate.}

\item{seed}{This gives you the same result again.}
}
\description{
This uses item model 4 and 5 to create number series items.
}
\details{
Comprehension of abstract object representation
i.e combination of:same number / same letter
2 simple linear (without arithmetic) 1 2 3 / 10 20 30 combine to form a number series item.
}
\examples{
\dontrun{

nmFive(items=5, seed=5)

}
}
\author{
Aiden Loe and Filip Simonfy
}

