% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel4_3.R
\name{nmIrregular}
\alias{nmIrregular}
\title{Item Model 13}
\usage{
nmIrregular(items)
}
\arguments{
\item{items}{Generate a random mix of items.}
}
\description{
This uses item model 13 to create number series items.
}
\details{
This function creates number series that is a irregular combination of sequences a b b a b b a ... Only the addition and substraction arithmetic operators are used to create the number series items.
}
\examples{
\dontrun{

#Draws 5 items randomly.
nmIrregular(5)

}
}
\author{
Aiden Loe and Filip Simonfy
}

