% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_Model_9.R
\name{nmNeighbour}
\alias{nmNeighbour}
\title{Item Model 9}
\usage{
nmNeighbour(cat, n, items, arith)
}
\arguments{
\item{cat}{Number of categorical groups per question.}

\item{n}{The differences between the pair of objects}

\item{items}{The number of items you want to generate.}

\item{arith}{The arithmetic operator of your choice ("add","substr","multi","div").}
}
\description{
This uses item model 9 to create number series items.
}
\details{
This is based on the categorical / pattern recognition rule. Neighbouring pairs or triads of objects are related, includes arithmetic operations.
}
\examples{
\dontrun{

nmNeighbour(2,4,2, "add")

}
}
\author{
Aiden Loe and Filip Simonfy
}

