% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel1_1.R
\name{nmOne}
\alias{nmOne}
\title{Item Model 1}
\usage{
nmOne(x, items)
}
\arguments{
\item{x}{To get the length value}

\item{items}{The number of items you have to generate.}
}
\description{
This uses item model 1 to create number series items.
}
\details{
counting the alphabets representation.
}
\examples{
\dontrun{

nmOne(3,26)

}
}
\author{
Aiden Loe and Filip Simonfy, \email{bsl28@cam.ac.uk}
}

