% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel3_1.R
\name{nmParaSeq}
\alias{nmParaSeq}
\title{Item Model 6}
\usage{
nmParaSeq(arith = "add", n = 1, items = 2)
}
\arguments{
\item{arith}{Select the arithmetric operator of choice ("add","multi", "sub", "div").}

\item{n}{Value you want use the arithmetic operator on.}

\item{items}{Generate a random mix of items.}
}
\description{
i.e combination of: sequences x y x y x y x y.
The selected arithmetic and the operator will remind the same between the combined sequences.
}
\details{
2 linear sequences (With use of the same arithmetic operator)
}
\examples{
\dontrun{

set.seed(1)
nmParaSeq(arithmetic="add",n=2,5)

}
}
\author{
Aiden Loe and Filip Simonfy
}

