% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel3_2_3.R
\name{nmProgress}
\alias{nmProgress}
\title{Item Model 8}
\usage{
nmProgress(vOne, vTwo, items, seed, logic = "one", random = FALSE)
}
\arguments{
\item{vOne}{The first value in the complex coefficient (x). Can be a sequence of values or a specific value.}

\item{vTwo}{The second value in the complex coefficient (y). Can be a sequence of values or a specific value.}

\item{items}{Generate a random mix of items.}

\item{seed}{To get the same random sampling of items}

\item{logic}{"one" or "two"}

\item{random}{If random=FALSE, the items will follow in sequential order.}
}
\description{
This function creates number series that is a combination of Arithmetic + Linear + Complex coefficient. The minimum number that will be generated is 10, given multi and add are both 1.
}
\details{
Arithmetic + Linear + Complex coefficient.
First logic of complex coefficient = i*x+y.
Second logic of complex coefficient = (i+x)*y.
}
\examples{
\dontrun{

#Draws 5 items randomly.
nmProgress(vOne=1,vTwo=3,items=5,seed=2,logic="one",random=TRUE)

# Calculates all combinations
# Items and seed arg is ignored.
nmProgress(vOne=1:2,vTwo=1:3,items=5,seed=2,logic="one",random=FALSE)

}
}
\author{
Aiden Loe and Filip Simonfy
}

