% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel3_4.R
\name{nmRatios}
\alias{nmRatios}
\title{Ratios}
\usage{
nmRatios(items, combo = "one", seed)
}
\arguments{
\item{items}{Generate a random mix of items.}

\item{combo}{There are four different type of combination.}

\item{seed}{To get the same random sampling of items}
}
\description{
This function creates number series that is a combination between relationships of objects and arithmetic (ratios). The maximum number to be generated is 80. There are four different types of model combination.
}
\details{
Relationships between objects + arithmetic (ratios).
}
\examples{
\dontrun{

# Combo 1. a : b ; c : d ; e : f
# Combo 2. a*x = b; c*x = d; e*x = f
# Combo 3. a : b : c = d : e : f = g : h : i
# Combo 4. a+b = x, c+d = x, e+f = x

# Draws 5 items
nmRatios(items = 5,combo ="one", 1)

}
}
\author{
Aiden Loe and Filip Simonfy
}

