% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel2_2.R
\name{nmRepresent}
\alias{nmRepresent}
\title{Letter Representation}
\usage{
nmRepresent(numLetter, value, items, random = FALSE)
}
\arguments{
\item{numLetter}{Number of Letters you want.}

\item{value}{Value + numLetter gives you back the total number of letters.}

\item{items}{Number of items to generate.}

\item{random}{If random=FALSE, the items will follow in sequential order.}
}
\description{
This function tells you the count of a specific letter.
}
\details{
Returns the number representation which is the count of the specific letters.
}
\examples{
\dontrun{

nmRepresent(numLetter= 2 ,value=3, items=10, random=FALSE)

}
}
\author{
Aiden Loe
}

