% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item_Model_7.R
\name{nmSeven}
\alias{nmSeven}
\title{Item Model 7}
\usage{
nmSeven(items)
}
\arguments{
\item{items}{Number of items to generate.}
}
\description{
This uses item model 7 to create number series items.
}
\details{
This model uses the addition and substraction (Arithmetic) operator, Linear pattern and Progressive coefficient to create the number series.
}
\examples{
\dontrun{

nmSeven(items=3)

}
}
\author{
Aiden Loe and Filip Simonfy
}

