% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nModel3_1.R
\name{nmSix}
\alias{nmSix}
\title{Item Model 6}
\usage{
nmSix(arith = "add", n = 1, items = 2)
}
\arguments{
\item{arith}{Select the arithmetric operator of choice ("add","multi", "sub", "div").}

\item{n}{Value you want use the arithmetic operator on.}

\item{items}{Generate a random mix of items.}
}
\description{
This uses item model 6 to create number series items.
}
\details{
2 linear sequences (With use of the same arithmetic operator).
Combination of: sequences x y x y x y x y.
The selected arithmetic and the operator will remind the same between the combined sequences.
}
\examples{
\dontrun{

set.seed(1)
nmSix(arithmetic="add",n=2,5)

}
}
\author{
Aiden Loe and Filip Simonfy
}

