% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nModel3_5.R
\name{nmTen}
\alias{nmTen}
\title{Item Model 10}
\usage{
nmTen(items)
}
\arguments{
\item{items}{Number of items to generate.}
}
\description{
This uses item model 10 to create number series items.
}
\details{
This function creates Fibonacci sequences. The maximum number to be generated is 15 items.
}
\examples{
\dontrun{

nmTen(items=3)

}
}
\author{
Aiden Loe and Filip Simonfy
}

