% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel2_1.R
\name{nmThree}
\alias{nmThree}
\title{Item Model 3 (Arithmetic)}
\usage{
nmThree(items, n, arithmetic)
}
\arguments{
\item{items}{The number of items you want to generate}

\item{n}{Value you want use the arithmetic operator on}

\item{arithmetic}{Use either 'add', 'substr', 'multi', 'div'.}
}
\description{
This uses item model 3 to create number series items.
}
\details{
This is based on linearity rule / pattern recognition rule and also arithmetic operators. Currently it only displays up to a series of 9.
}
\examples{
\dontrun{

nmThree(items=4,n=2,arithmetic="add")

}
}
\author{
Aiden Loe and Filip Simonfy
}

