% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeLength.R
\name{nodeLength}
\alias{nodeLength}
\title{Node Length}
\usage{
nodeLength(rank)
}
\arguments{
\item{R}{This is the Rank of the maze.}
}
\details{
This function tells us the nodes on the longest row given a rank.
This needs to have a rank value of greater than 1.
}
\author{
Aiden Loe
}

