% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{solution}
\alias{solution}
\title{solution}
\usage{
solution(rank, nodePosition)
}
\arguments{
\item{rank}{This is the Rank of the maze}

\item{nodePosition}{Tells you all the position of the black dots.}
}
\description{
This function tells us the solution for all series of routes including both black and non black dots.
}
\details{
This function tells us the solution for all series of routes including both black and non black dots.
}
\examples{
rank <- 3
a <- colourNodePosition(rank=3,satPercent=0.5,seed=1)
solution(rank,a)
}
\author{
Aiden Loe and Maria Sanchez
}

