% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookUniqueNodePositions.R
\name{uniqueNodePositions}
\alias{uniqueNodePositions}
\title{Unique Node Positions}
\usage{
uniqueNodePositions(rank, satPercent, seed)
}
\arguments{
\item{rank}{This is the Rank of the maze.}

\item{satPercent}{Percentage of saturation.}

\item{seed}{Returns a unique node distribution specific to the local computer.}
}
\description{
The function returns a black nodes distribution with a single unique solution and seed number
}
\details{
The function returns a black nodes distribution with a single unique solution and seed number
}
\examples{
\dontrun{

rank <- 5
satPercent <- 0.5

#Searches for just one unique solution
justOne <- genUniqueSolution(rank,satPercent,11)

#Black nodes distribution for single unique solution + seed number
uniqueNodePositions(rank,satPercent,justOne)

colourNodePosition(rank,satPercent,seed=justOne)
}
}
\author{
Aiden Loe and Maria Sanchez
}

