% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genuniquesolution.R
\name{genUniqueSolution}
\alias{genUniqueSolution}
\title{genUniqueSolution}
\usage{
genUniqueSolution(rank, satPercent, seed)
}
\arguments{
\item{rank}{This is the Rank of the maze.}

\item{satPercent}{Percentage of saturation.}

\item{seed}{Returns the seed that only gives one unique solution for a given rank and saturation.}
}
\description{
The generate unique solution function searches for the SEED that returns only one unique solution for a given rank and saturation.
}
\details{
The generate unique solution function searches for the SEED that returns only one unique solution for a given rank and saturation. This might be computationally intensive as the maze size increases. The seed is necessary so that the algorithm does not always begin from the smallest value. Based on the seed value, it will search for the next biggest that returns 1 unique solution.
}
\examples{

rank <- 5
satPercent <- 0.5
seed <- 1
#Searches for just one unique solution
justOne <- genUniqueSolution(rank=rank,satPercent=satPercent,seed=seed)


}
\author{
Aiden Loe and Maria Sanchez
}

