% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxScore.R
\name{maxScore}
\alias{maxScore}
\title{Maximum Score}
\usage{
maxScore(rank, nodePosition)
}
\arguments{
\item{rank}{This is the Rank of the maze.}
}
\details{
The maxScore function returns the maximum score for a given rank and a given node Position. You need to use the colour node position function first.
}
\author{
Aiden Loe
}

