% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mazeGen.R
\name{mazeGen}
\alias{mazeGen}
\title{mazeGen: A package for generating number series items.}
\description{
The mazeGen package provides 14 item models for generating number series items
}
\details{
Still need to make changes here.
}
\section{Item model 1}{

This number series counts the alphabets representation.\cr
\code{\link{nmOne}}
}

