% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nModel2_1.R
\name{nmAdd}
\alias{nmAdd}
\title{Item Model 3 (Addition)}
\usage{
nmAdd(length, items)
}
\arguments{
\item{length}{This is the length of each item}

\item{items}{The number of items you want to generate}
}
\description{
This uses item model 3 to create number series items.
}
\details{
This is based on linearity rule / pattern recognition rule.
}
\examples{
\dontrun{

nmAdd(4,5)

}
}
\author{
Aiden Loe and Filip Simonfy
}

