% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel2_1.R
\name{nmAlpha}
\alias{nmAlpha}
\title{Item Model 3 (Alpha)}
\usage{
nmAlpha(items, reverse = FALSE)
}
\arguments{
\item{items}{The number of items you want to generate}

\item{reverse}{Changing reverse to TRUE will flip to direction of the progressive alphabets}
}
\description{
Item Model 3 (Alpha)
}
\details{
This is based on linearity rule / pattern recognition rule
This is for alphabets and is restricted to 21 items.
}
\examples{
\dontrun{

nmAlpha(3,reverse=TRUE)

}
}
\author{
Aiden Loe and Filip Simonfy
}

