% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel1_3.R
\name{nmCat}
\alias{nmCat}
\title{Group Categorisation}
\usage{
nmCat(cat, items, random = FALSE)
}
\arguments{
\item{cat}{Length of categorical groups per question.}

\item{items}{The number of items you want to generate.}

\item{random}{To randomise the position of the numeric values.}
}
\description{
This is based on the categorical / pattern recognition rule.
}
\details{
Identification of the number series follows a group pattern recognition rule.
}
\examples{
\dontrun{

nmCat(cat=2,items=4,random=FALSE)

}
}
\author{
Aiden Loe and Filip Simonfy
}

