% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel4_1.R
\name{nmEleven}
\alias{nmEleven}
\title{Item Model 11}
\usage{
nmEleven(items, logic = "one", n = 2, arith = "add")
}
\arguments{
\item{items}{Generate a random mix of items.}

\item{logic}{The combination of sequences follow two logic.}

\item{n}{The value that the arithmetic operator uses to calculate the next value}

\item{arith}{The arithmetic operator of your choice ("add","substr","multi","div").}
}
\description{
This uses item model 11 to create number series items.
}
\details{
The number series items are a combination of Arithmetic, linear sequence and progressive coefficient.
First logic is combining sequences x y x y x y x y = one simple, one progressive.
Second logic is combining sequences x y x y x y x y = two progressive. The minimum number of items that will be generated is 2.
}
\examples{
\dontrun{

#Draws 10 items randomly.
nmEleven(10,logic="one", n=2,arith="add")

}
}
\author{
Aiden Loe and Filip Simonfy
}

