% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel2_2.R
\name{nmFour}
\alias{nmFour}
\title{Item Model 4}
\usage{
nmFour(numLetter = 2, value = 3, items = 3, random = FALSE)
}
\arguments{
\item{numLetter}{Number of Letters you want.}

\item{value}{Value + numLetter gives you back the total number of letters.}

\item{items}{Number of items to generate.}

\item{random}{If random=FALSE, the items will follow in sequential order.}
}
\description{
This function tells you the count of a specific letter.
}
\details{
The core feature of sequences is the numeric representation of the object count.  The sequences consist of homogeneous groups of letters followed by a number representing the amount of the group elements.
}
\examples{
\dontrun{

nmFour(numLetter= 2 ,value=3, items=10, random=FALSE)

}
}
\author{
Aiden Loe
}

