% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel4_3.R
\name{nmIrregular}
\alias{nmIrregular}
\title{Item Model 13}
\usage{
nmIrregular(items)
}
\arguments{
\item{items}{Generate a random mix of items.}
}
\description{
This function creates number series that is a irregular combination of sequences a b b a b b a ... The values are big because the limits of the div and multiplication are allowed high. Also, we can limit the values of add and str by changing the value 99 to something smaller in the code.
}
\details{
# # Irregular combination of sequences a b b a b b a ...
}
\examples{
\dontrun{

#Draws 5 items randomly.
nmIrregular(5)

}
}
\author{
Aiden Loe and Filip Simonfy
}

