% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel4_4.R
\name{nmSeqRatio}
\alias{nmSeqRatio}
\title{Item Model 14}
\usage{
nmSeqRatio(items = 10, seed = 1)
}
\arguments{
\item{items}{Generate a random mix of items.}

\item{seed}{Generate a fixed ratio set but not final item set.}
}
\description{
This uses item model 14 to create number series items.
}
\details{
Combination of sequences and ratios. This function creates number series creates a combination of sequences and ratios. The seed generates a fixed ratio set to be combined with sequence but does not fix the final item set. Only the addition and substraction arimethic operators are used to generate the number series items.
}
\examples{
\dontrun{

#Draws 10 items randomly.
nmSeqRatio(10,5)

}
}
\author{
Aiden Loe and Filip Simonfy
}

