% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_model_7.r
\name{nmSeven}
\alias{nmSeven}
\title{Item Model 7}
\usage{
nmSeven(items)
}
\arguments{
\item{items}{Number of items to generate.}
}
\description{
This uses item model 7 to create number series items - Identification of progressively evolving coefficients of change.
}
\details{
Non-linear progressive sequences which require a higher level of abstraction; the coefficient of change between two neighbouring elements is not invariable and its elements form a sequence. The coefficient sequences correspond to items from Item Families 1 and 3. Example: The coefficient of change between each pair of neighbouring elements in the sequence increases by 1. (2 4 7 11 16 (22))
}
\examples{
\dontrun{

nmSeven(items=3)

}
}
\author{
Aiden Loe and Filip Simonfy
}

