% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Item_Model_7.R
\name{nmSeven}
\alias{nmSeven}
\title{Item Model 7}
\usage{
nmSeven(items)
}
\arguments{
\item{items}{Number of items to generate.}
}
\description{
Arithmetic + Linear + Progressive coefficient (i.e. coefficient = sequence)
2 4 7 11 16 ?
define starting point - e.g. each number 1:100
define coefficient sequence - first element and progression
element = first element + sum of the sequence at the position of the element
}
\details{
This function creates number series items a hypothesized cognitive model.
}
\examples{
\dontrun{

nmSeven(items=3)

}
}
\author{
Aiden Loe and Filip Simonfy
}

