% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nModel4_3.R
\name{nmThirteen}
\alias{nmThirteen}
\title{Item Model 13}
\usage{
nmThirteen(items)
}
\arguments{
\item{items}{Generate a random mix of items.}
}
\description{
This uses item model 13 to create number series items - Identification of unevenly ordered sub-sequences
}
\details{
This function creates number series that is a irregular combination of sequences a b b a b b a ... Only the addition and substraction arithmetic operators are used to create the number series items.
}
\examples{
\dontrun{

#Draws 5 items randomly.
nmThirteen(100)

}
}
\author{
Aiden Loe and Filip Simonfy
}
