% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel4_2.R
\name{nmTwelve}
\alias{nmTwelve}
\title{Item Model 12}
\usage{
nmTwelve(items = 1, fun1 = "add", fun2 = "add")
}
\arguments{
\item{items}{Generate a random mix of items.}

\item{fun1}{The argument decides the arithmetic to be employed for Neighbouring objects. There are two arithmetic: add, substr.}

\item{fun2}{The argument decides the arithmetic to be employed for the two values between the grouped objects.  There are two arithmetic: add, substr.}
}
\description{
This function creates number series that is a combination of Neighbouring objects + 2-sequence coefficient. Multiplication and Division is removed since the calculated value is too big.
}
\details{
# # Neighbouring objects + 2-sequence coefficient.
}
\examples{
\dontrun{

#Draws 5 items randomly.
nmTwelve(items=5, fun1 = "add", fun2= "add")

}
}
\author{
Aiden Loe and Filip Simonfy
}

